vaccel-client: Add shared object resource functionality (#3)
- add 32bit build cast
- add shared object resource functionality
- update bindings pointer

+++ vaccel-client/src/tf_model.rs
@@ -119,7 +119,7 @@ impl VsockClient {
                 ffi::vaccel_tf_tensor_set_data(
                     tensor,
                     data.as_ptr() as *mut std::ffi::c_void,
-                    data.len() as u64,
+                    data.len() as usize,
                 );
 
                 std::mem::forget(data);
